// packages/component-library/src/core/DataPointsTable/schemas/DeviceProfileSchema.ts
/* eslint-disable */
/* prettier-ignore-file */
// @ts-nocheck
import { z } from 'zod';
import { Dnp3ClientResourceSchema } from './DeviceResources/Dnp3ClientResourceSchema';
import { Dnp3ServerResourceSchema } from './DeviceResources/Dnp3ServerResourceSchema';
import { ModbusClientResourceSchema } from './DeviceResources/ModbusClientResourceSchema';
import { Iec60870104ClientResourceSchema } from './DeviceResources/Iec60870104ClientResourceSchema';
import { Iec60870104ServerResourceSchema } from './DeviceResources/Iec60870104ServerResourceSchema';
import { Iec61850MmsClientResourceSchema } from './DeviceResources/Iec61850MmsClientResourceSchema';

export const DeviceResourceSchema = z.union([
    Dnp3ClientResourceSchema,
    Dnp3ServerResourceSchema,
    ModbusClientResourceSchema,
    Iec60870104ClientResourceSchema,
    Iec60870104ServerResourceSchema,
    Iec61850MmsClientResourceSchema,
]);

export type DeviceResourceSchemaType = typeof DeviceResourceSchema;

export type DeviceResource = z.infer<typeof DeviceResourceSchema>;

export const DeviceProfileSchema = z.object({
    // The device name is required and needs to be unique across EdgeX.
    // TODO: Determine if we want an async refinement here to query EdgeX for existence.
    name: z
        .string('Profile name is required')
        .min(1, 'Profile name cannot be empty')
        .regex(/^[a-zA-Z0-9-_]+$/, 'Profile name can only contain alphanumeric characters, hyphens, and underscores'),
    // The id, created, modified fields are generated by EdgeX and are thus optional.
    id: z.string().optional(),
    created: z.number().optional(),
    modified: z.number().optional(),
    // The description, labels, manufacturer, and model are optional metadata fields.
    description: z.string().optional(),
    labels: z.array(z.string()).optional(),
    manufacturer: z.string().optional(),
    model: z.string().optional(),
    // The device resources and commands are required, but can be empty arrays.
    deviceResources: z.array(DeviceResourceSchema),
    // TODO: Implement device commands schema if that ever becomes relevant.
    deviceCommands: z.array(z.any()),
});
export type DeviceProfile = z.infer<typeof DeviceProfileSchema>;